requirements-api.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
python-dotenv>=1.2.0,<2.0.0
pandas<3
numpy>=2.0.0,<3.0.0
scikit-learn>=1.8.0,<2.0.0
xgboost>=3.0.0,<4.0.0
